requirements.txt
flask
flask-mysql-connector
flask-bcrypt
flask-login
python-dotenv
